move information outside buttons, cleaning
