+++ packages/atlas/src/components/Information/Information.styles.ts
