+++ packages/atlas/src/components/Information/Information.tsx
