+++ packages/atlas/src/components/Information/index.ts
