--- packages/atlas/src/components/_buttons/InformationButton/index.ts
@@ -1,1 +1,0 @@
-export * from './InformationButton'
